Small bug fixes and safety throws

+++ scripts/resetDatabase.ts
@@ -5,6 +5,10 @@ import config from "../knexfile";
 const db = knex(config.production);
 
 (async () => {
+  if (JSON.stringify(config.production).includes(".cockroachlabs.cloud")) {
+    throw new Error("Don't reset the production database, please.");
+  }
+
   console.log("Dropping existing database...");
   await db.raw('DROP SCHEMA IF EXISTS "public" CASCADE');
